requirements.txt
python-telegram-bot==22.*
SQLAlchemy==2.*
sentry-sdk==2.*
python-dotenv==1.*
alembic==1.*
pytest==8.*
